Rename upload_json to s3_control
s3와 관련된 작업을 하나의 스크립트에서 처리해주는 것이 더 효율적이라 생각해 s3와
관련된 일을 하는 함수들을 병합했습니다. upload는 스크래핑과 같이 동작하지만, send_model은
스케줄러로 설정해줘야 합니다.
